Modified the validations part 1

+++ Partnerly/Infrastructure/Services/UserService.cs
@@ -1,19 +1,29 @@
-﻿using Partnerly.Infrastructure.Interfaces;
+﻿using Partnerly.Descriptors.Messages;
+using Partnerly.Helpers;
+using Partnerly.Infrastructure.Interfaces;
 using Partnerly.Models;
+using System.ComponentModel.DataAnnotations;
 
 namespace Partnerly.Infrastructure.Services
 {
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepo;
+        private readonly IPermissionService _permissionService;
+        private readonly ICurrentUserService _currentUser;
 
-        public UserService(IUserRepository userRepo)
+        public UserService(IUserRepository userRepo, IPermissionService permissionService, ICurrentUserService currentUser)
         {
             _userRepo = userRepo;
+            _permissionService = permissionService;
+            _currentUser = currentUser;
         }
 
-        public async Task<User> CreateUserAsync(User user)
+        public async Task<User?> CreateUserAsync(User? user)
         {
+            if (user == null)
+                return null;
+
             var newUser = new User { Id = Guid.NewGuid() };
             newUser.Email = user.Email;
             newUser.Phone = user.Phone;
@@ -29,6 +39,12 @@ namespace Partnerly.Infrastructure.Services
             newUser.CreatedBy = newUser.Id;
             newUser.UpdatedBy = newUser.Id;
 
+            string? result = ValidationHelper.ValidateEntityRequiredFields(newUser, out bool isValid);
+            if (!isValid)
+            {
+                throw new ValidationException(String.Format(ErrorMessages.RequiredFieldsValidationFailed, result));
+            }
+
             await _userRepo.AddAsync(newUser);
             await _userRepo.SaveChangesAsync();
             return newUser;
@@ -37,20 +53,40 @@ namespace Partnerly.Infrastructure.Services
         public async Task<User?> GetUserByEmailAsync(string email) =>
             await _userRepo.GetByEmailAsync(email);
 
-        public async Task<IEnumerable<User>> GetAllUsersAsync() =>
+        public async Task<User?> GetUserByIDAsync(Guid id) =>
+            await _userRepo.GetByIdAsync(id);
+
+        public async Task<IEnumerable<User?>> GetAllUsersAsync() =>
             await _userRepo.GetAllAsync();
 
-        public async Task UpdateUserAsync(User user)
+        public async Task UpdateUserAsync(User? user)
         {
+            if (_currentUser.UserId == null || user == null)
+                return;
+
+            if (!await _permissionService.CanUpdateAsync(_currentUser.UserId, user))
+                throw new UnauthorizedAccessException(ErrorMessages.NoPermissionForThisAction);
+
+            string? result = ValidationHelper.ValidateEntityRequiredFields(user, out bool isValid);
+            
+            if (!isValid)
+                throw new ValidationException(String.Format(ErrorMessages.RequiredFieldsValidationFailed, result));
+
             _userRepo.Update(user);
             await _userRepo.SaveChangesAsync();
         }
 
-        public async Task DeleteUserAsync(Guid id)
+        public async Task DeleteUserAsync(Guid? id)
         {
-            var user = await _userRepo.GetByIdAsync(id);
+            if (id == null || _currentUser.UserId == null)
+                return;
+
+            var user = await _userRepo.GetByIdAsync((Guid)id);
             if (user != null)
             {
+                if (!await _permissionService.CanDeleteAsync(_currentUser.UserId, user))
+                    throw new UnauthorizedAccessException(ErrorMessages.NoPermissionForThisAction);
+
                 _userRepo.Delete(user);
                 await _userRepo.SaveChangesAsync();
             }
